src/requirements.txt
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.5.0
async-timeout==4.0.3
asyncpg==0.29.0
black==24.8.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
fastapi==0.115.0
greenlet==3.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-env==1.1.5
python-dotenv==1.0.1
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6
watchfiles==0.24.0
websockets==13.0.1
